Factor out decoders into separate files
